Reorganize form modal code
