--- projects/packages/forms/src/blocks/contact-form/components/hooks/index.js
@@ -1,2 +1,0 @@
-export { useIntegrationStatus } from './useIntegrationStatus';
-export { usePluginInstallation } from './usePluginInstallation';
