+++ projects/packages/forms/src/blocks/contact-form/components/jetpack-integrations-modal/creative-mail-card.js
@@ -1,3 +1,4 @@
+import { store as blockEditorStore } from '@wordpress/block-editor';
 import { createBlock } from '@wordpress/blocks';
 import { ExternalLink, ToggleControl } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
@@ -7,9 +8,9 @@ import IntegrationCard from './integration-card';
 const CreativeMailCard = ( { isExpanded, onToggle, data, refreshStatus } ) => {
 	const { settingsUrl = '' } = data || {};
 
-	const selectedBlock = useSelect( select => select( 'core/block-editor' ).getSelectedBlock(), [] );
+	const selectedBlock = useSelect( select => select( blockEditorStore ).getSelectedBlock(), [] );
 
-	const { insertBlock, removeBlock } = useDispatch( 'core/block-editor' );
+	const { insertBlock, removeBlock } = useDispatch( blockEditorStore );
 
 	const hasEmailBlock = selectedBlock?.innerBlocks?.some(
 		( { name } ) => name === 'jetpack/field-email'
